Force $methods to be an array
With PHP7.2, calling count on a non Countable causes a warning.

This bundle calls count() on the RateLimit->getMethods() return value and will fail if it is a string. We had a use in one of our apps where it was a string, which was incorrect, but this brings the methods property on this annotation into the same behaviour as the rest of Symfony's handling of methods.